puts all the styled-components in a seperate jsx file
